Agrego cada funcion en un archivo distinto, agrego documentacion y correji algunos errores, asi mismo, es la ultima y definitiva version del proyecto
