+++ proyectos/micro-sist-de-arch-multihilos/TenorioJesus/.gitignore
@@ -1,0 +1,6 @@
+.img
+__pycache__
+.org
+.png
+.jpg
+.docx
